requirements.txt
maafw
readerwriterlock